Migrate to sovereign metal (#97)
* compiling

* wip

* remove buffer ref

* remove unused type aliases from metal backend

* remove TextureRef type alias

* remove Buffer type alias

* remove CommandBuffer type alias

* remove ComputeCommandEncoder type alias

* remove ComputePipelineState type alias

* remove unused CommandQueue type alias

* remove Device type alias

* remove unused Heap type alias

* remove unused Library type alias

* remove rest typealiases

* use ndarray to speedup attention kernel test

* update quant matmul

* remove legacy constant values

* comment out flaky test

* remove label ext

* remove unused reflection

* make condition safe

* use convenience encoder fubcs

* use published metal

+++ crates/uzu/src/backends/metal/kernel/matmul/gemv/pipeline_configuration.rs
@@ -1,4 +1,4 @@
-use metal::MTLSize;
+use crate::backends::metal::MTLSize;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct PipelineConfiguration {
@@ -31,8 +31,8 @@ impl PipelineConfiguration {
     pub fn threads_per_threadgroup(&self) -> MTLSize {
         MTLSize::new(
             32,
-            self.threadgroup_cols as u64,
-            self.threadgroup_rows as u64,
+            self.threadgroup_cols as usize,
+            self.threadgroup_rows as usize,
         )
     }
 }
