+++ crates/uzu/src/backends/metal/kernel/mlp_fused/gemv/pipeline_configuration.rs
@@ -1,6 +1,4 @@
-use metal::MTLSize;
-
-use crate::backends::metal::kernel::mlp::MlpActivationType;
+use crate::backends::metal::{MTLSize, kernel::mlp::MlpActivationType};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct PipelineConfiguration {
@@ -23,8 +21,8 @@ impl PipelineConfiguration {
     pub fn threads_per_threadgroup(&self) -> MTLSize {
         MTLSize::new(
             32,
-            self.threadgroup_cols as u64,
-            self.threadgroup_rows as u64,
+            self.threadgroup_cols as usize,
+            self.threadgroup_rows as usize,
         )
     }
 }
